Add "old" realtime .xml files to new format.
parse.py now detects the version of the script it's reading and
responds appropriately: the old .xml format will keep the static
f107d described in the .xml header and assign the observed Kp to the
24hr Kp average.
